Renamed module for dockerized keycloak.
Calling it by a version number is unsustainable
