+++ security-oauth2/build.gradle.kts
@@ -26,7 +26,7 @@ dependencies {
     testImplementation(projects.micronautSecurityJwt)
     testImplementation(projects.testSuiteUtils)
     testImplementation(projects.testSuiteUtilsSecurity)
-    testImplementation(projects.testSuiteKeycloak16)
+    testImplementation(projects.testSuiteKeycloakDocker)
     testImplementation(mnLogging.logback.classic)
     testImplementation(libs.system.stubs.core)
 }
