+++ security-oauth2/src/test/groovy/io/micronaut/security/oauth2/endpoint/authorization/request/pkce/OpenIdAuthorizationPkceSpec.groovy
@@ -13,7 +13,7 @@ import io.micronaut.security.oauth2.client.OpenIdClient
 import io.micronaut.security.oauth2.endpoint.authorization.pkce.S256PkceGenerator
 import io.micronaut.security.oauth2.routes.OauthController
 import io.micronaut.security.testutils.EmbeddedServerSpecification
-import io.micronaut.security.oauth2.keycloak.v16.Keycloak
+import io.micronaut.security.oauth2.keycloak.docker.Keycloak
 import org.testcontainers.DockerClientFactory
 import spock.lang.IgnoreIf
 import spock.lang.Requires
