+++ security-oauth2/src/test/groovy/io/micronaut/security/oauth2/endpoint/authorization/request/pkce/OpenIdAuthorizationRedirectPkceSpec.groovy
@@ -19,10 +19,7 @@ import io.micronaut.security.oauth2.endpoint.token.response.OauthAuthenticationM
 import io.micronaut.security.oauth2.endpoint.token.response.TokenResponse
 import io.micronaut.security.oauth2.routes.OauthController
 import io.micronaut.security.testutils.EmbeddedServerSpecification
-import io.micronaut.security.oauth2.keycloak.v16.Keycloak
-import jakarta.inject.Named
-import jakarta.inject.Singleton
-import org.reactivestreams.Publisher
+import io.micronaut.security.oauth2.keycloak.docker.Keycloak
 import jakarta.inject.Named
 import jakarta.inject.Singleton
 import org.reactivestreams.Publisher
