+++ settings.gradle
@@ -26,7 +26,7 @@ include "test-suite-security-service"
 include "test-suite-http"
 include "test-suite-geb"
 include "test-suite-keycloak"
-include "test-suite-keycloak-16"
+include "test-suite-keycloak-docker"
 include "test-suite-tomcat"
 include "test-suite-tomcat-groovy"
 include "test-suite-utils"
