+++ test-suite-geb/src/test/groovy/io/micronaut/security/utils/KeycloakCleanup.groovy
@@ -1,6 +1,6 @@
 package io.micronaut.security.utils
 
-import io.micronaut.security.oauth2.keycloak.v16.Keycloak
+import io.micronaut.security.oauth2.keycloak.docker.Keycloak
 import org.spockframework.runtime.extension.IGlobalExtension
 
 class KeycloakCleanup implements IGlobalExtension {
