+++ test-suite-keycloak-docker/build.gradle
